refactor the coundown component
