+++ src/css/override.css
@@ -9,7 +9,7 @@
 }
 
 .slick-list, .slick-track {
-    height: 320px;
+    /* height: 320px; */
     display: flex;
 }
 
@@ -70,6 +70,6 @@
     height: 20px;
 }
 
-#section-items .container .row .slick-slider {
+/* #section-items .container .row .slick-slider {
     height: 516px;
-}
+} */
